pyproject.toml
[project]
name = "media-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "b2sdk>=2.9.4",
    "fastmcp>=2.8.1",
    "ffmpeg-python>=0.2.0",
    "pytest>=8.4.1",
    "requests>=2.32.4",
    "uuid>=1.30",
]
